requirements.txt
Flask
jwt
bcrypt
datetime
pymongo
python-dotenv
beautifulsoup4
jinja2
requests
